Closes https://github.com/datadryad/dryad-product-roadmap/issues/4045
