+++ app/javascript/react/components/MetadataEntry/Publication/Publication.jsx
@@ -1,8 +1,27 @@
 import React, {useState, useEffect} from 'react';
 import axios from 'axios';
+import {titleCase} from '../../../../lib/title-case';
 import {showSavedMsg, showSavingMsg, formatSizeUnits} from '../../../../lib/utils';
 import PublicationForm from './PublicationForm';
 import Title from './Title';
+
+const copyTitle = (e) => {
+  const copyButton = e.currentTarget.firstElementChild;
+  const title = e.currentTarget.previousSibling.textContent;
+  navigator.clipboard.writeText(title).then(() => {
+    // Successful copy
+    copyButton.parentElement.setAttribute('title', 'Title copied');
+    copyButton.classList.remove('fa-paste');
+    copyButton.classList.add('fa-check');
+    copyButton.innerHTML = '<span class="screen-reader-only">Title copied</span>';
+    setTimeout(() => {
+      copyButton.parentElement.setAttribute('title', 'Copy title');
+      copyButton.classList.add('fa-paste');
+      copyButton.classList.remove('fa-check');
+      copyButton.innerHTML = '';
+    }, 2000);
+  });
+};
 
 export default function Publication({resource, setResource, maxSize}) {
   const subType = resource.resource_type.resource_type;
@@ -11,6 +30,7 @@ export default function Publication({resource, setResource, maxSize}) {
   const [showTitle, setShowTitle] = useState(false);
   const [importType, setImportType] = useState(resource.identifier.import_info);
   const [sponsored, setSponsored] = useState(false);
+  const [caseWarning, setCaseWarning] = useState(false);
   const [dupeWarning, setDupeWarning] = useState(false);
 
   const authenticity_token = document.querySelector("meta[name='csrf-token']")?.getAttribute('content');
@@ -63,12 +83,22 @@ export default function Publication({resource, setResource, maxSize}) {
       setShowTitle(true);
     }
     setSponsored(!!res.journal?.payment_plan_type && (manuscript_number || primary_article) ? res.journal.title : false);
-    if (res.title && !resource.identifier.process_date?.processing) {
-      axios.get(`/resources/${resource.id}/dupe_check.json`).then((data) => {
-        setDupeWarning(data.data?.[0]?.title || false);
-      });
-    } else {
-      setDupeWarning(false);
+    if (res.title) {
+      if (!resource.identifier.process_date?.processing) {
+        axios.get(`/resources/${resource.id}/dupe_check.json`).then((data) => {
+          setDupeWarning(data.data?.[0]?.title || false);
+        });
+      } else {
+        setDupeWarning(false);
+      }
+      if (
+        res.title === res.title.toUpperCase()
+        || (res.title.match(/\b[A-Z].*?\b/g)?.length > res.title.split(/\s/).length * 0.6 && res.title !== titleCase(res.title, {sentenceCase: true}))
+      ) {
+        setCaseWarning(true);
+      } else {
+        setCaseWarning(false);
+      }
     }
   }, [res]);
 
@@ -135,6 +165,28 @@ export default function Publication({resource, setResource, maxSize}) {
 
       {showTitle && (
         <Title key={res.title} resource={res} setResource={setRes} />
+      )}
+
+      {caseWarning && (
+        <div className="callout warn">
+          <p style={{fontSize: '.98rem'}}>Please correct your dataset title to sentence case, which could look like:</p>
+          <p><span>{titleCase(res.title, {sentenceCase: true})}</span>
+            <span
+              className="copy-icon"
+              role="button"
+              tabIndex="0"
+              aria-label="Copy title"
+              title="Copy title"
+              onClick={copyTitle}
+              onKeyDown={(e) => {
+                if (e.key === ' ' || e.key === 'Enter') {
+                  copyTitle(e);
+                }
+              }}
+            ><i className="fa fa-paste" role="status" />
+            </span>
+          </p>
+        </div>
       )}
 
       {dupeWarning && (
