+++ app/javascript/react/components/MetadataEntry/Publication/index.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import {titleCase} from '../../../../lib/title-case';
 
 export {default} from './Publication';
 export {default as PubPreview} from './PubPreview';
@@ -11,34 +10,10 @@ const nondescript = (t) => {
   return remainder.split(/\s/).length < 4;
 };
 
-const capitals = (t) => {
-  if (t === t.toUpperCase()) return 'All-caps titles are not allowed.';
-  if (t.match(/\b[A-Z].*?\b/g)?.length > t.split(/\s/).length * 0.6) return 'Sentence case is preferred for titles.';
-  return false;
-};
-
 const validPrimary = (r, t = 'primary_article') => {
   const p = r.related_identifiers.find((ri) => ri.work_type === t);
   if (!p) return false;
   return /^https:\/\/doi\.org\/10\.\d{4,9}\/[-._;()/:a-zA-Z0-9]+$/.test(p.related_identifier);
-};
-
-const copyTitle = (e) => {
-  const copyButton = e.currentTarget.firstElementChild;
-  const title = e.currentTarget.previousSibling.textContent;
-  navigator.clipboard.writeText(title).then(() => {
-    // Successful copy
-    copyButton.parentElement.setAttribute('title', 'Title copied');
-    copyButton.classList.remove('fa-paste');
-    copyButton.classList.add('fa-check');
-    copyButton.innerHTML = '<span class="screen-reader-only">Title copied</span>';
-    setTimeout(() => {
-      copyButton.parentElement.setAttribute('title', 'Copy title');
-      copyButton.classList.add('fa-paste');
-      copyButton.classList.remove('fa-check');
-      copyButton.innerHTML = '';
-    }, 2000);
-  });
 };
 
 export const publicationPass = (resource) => !!resource.identifier.import_info || !!resource.title;
@@ -46,44 +21,6 @@ export const publicationPass = (resource) => !!resource.identifier.import_info |
 export const publicationFail = (resource, review) => {
   const {import_info} = resource.identifier;
   if (!!import_info || review) {
-    if (resource.title) {
-      if (nondescript(resource.title)) {
-        return (
-          <p className="error-text" id="title_error">
-            Your dataset title is not specific to your dataset. Use a descriptive title so your data can be discovered.
-          </p>
-        );
-      }
-      if (capitals(resource.title)) {
-        return (
-          <>
-            <p className="error-text" id="title_error">
-              {capitals(resource.title)} Please correct your dataset title to sentence case, which could look like:
-            </p>
-            <div className="callout warn">
-              <p><span>{titleCase(resource.title, {sentenceCase: true})}</span>
-                <span
-                  className="copy-icon"
-                  role="button"
-                  tabIndex="0"
-                  aria-label="Copy title"
-                  title="Copy title"
-                  onClick={copyTitle}
-                  onKeyDown={(e) => {
-                    if (e.key === ' ' || e.key === 'Enter') {
-                      copyTitle(e);
-                    }
-                  }}
-                ><i className="fa fa-paste" role="status" />
-                </span>
-              </p>
-            </div>
-          </>
-        );
-      }
-    } else {
-      return <p className="error-text" id="title_error">Title is required</p>;
-    }
     const {publication_name, manuscript_number} = resource.resource_publication;
     const {publication_name: preprint_server} = resource.resource_preprint || {};
     if (['manuscript', 'published'].includes(import_info) && !publication_name) {
@@ -110,6 +47,24 @@ export const publicationFail = (resource, review) => {
       return (
         <p className="error-text" id="doi_error">A valid DOI for the preprint is required</p>
       );
+    }
+    if (resource.title) {
+      if (nondescript(resource.title)) {
+        return (
+          <p className="error-text" id="title_error">
+            Your dataset title is not specific to your dataset. Use a descriptive title so your data can be discovered.
+          </p>
+        );
+      }
+      if (resource.title === resource.title.toUpperCase()) {
+        return (
+          <p className="error-text" id="title_error">
+            All-caps titles are not allowed.
+          </p>
+        );
+      }
+    } else {
+      return <p className="error-text" id="title_error">Title is required</p>;
     }
   }
   return false;
