+++ package.json
@@ -74,6 +74,7 @@
     "axios": "^1.8.2",
     "babel-loader": "8",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "brill": "^3.1.0",
     "compression-webpack-plugin": "9",
     "copy-webpack-plugin": "^11.0.0",
     "copyfiles": "^2.4.1",
@@ -83,7 +84,6 @@
     "draft-js": "^0.11.7",
     "drag-on-drop": "^3.6.1",
     "evaporate": "^2.1.4",
-    "fast-tag-pos": "^2.0.0",
     "formik": "^2.2.9",
     "html-react-parser": "^1.4.14",
     "identity-obj-proxy": "^3.0.0",
